refactor: renamed AddAITheoryBlockModal
