+++ frontend/src/lib/components/AddAITheoryBlockModal.svelte
